% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_cov_data}
\alias{get_cov_data}
\title{Get datasets from Vancouver Open Data Portal}
\usage{
get_cov_data(dataset_id, format = c("csv", "geojson"), where = NULL,
  apikey = getOption("VancouverOpenDataApiKey"), rows = NULL,
  refresh = FALSE)
}
\arguments{
\item{dataset_id}{Dataset id from the Vancouver Open Data catalogue}

\item{format}{`csv` or `geojson` are supported at this time (default `csv`)}

\item{where}{Query parameter to filter data (default `NULL` no filter)}

\item{apikey}{Vancouver Open Data API key, default `getOption("VancouverOpenDataApiKey")`}

\item{rows}{Maximum number of rows to return (default `NULL` returns all rows)}

\item{refresh}{refresh cached data, default `FALSE``}
}
\description{
Get datasets from Vancouver Open Data Portal
}
