% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_cov_metadata}
\alias{get_cov_metadata}
\title{Get metadata for CoV open data dataset}
\usage{
get_cov_metadata(
  dataset_id,
  apikey = getOption("VancouverOpenDataApiKey"),
  refresh = FALSE
)
}
\arguments{
\item{dataset_id}{the CoV open data dataset id}

\item{apikey}{the CoV open data API key, optional}

\item{refresh}{refresh cached data, default `FALSE``}
}
\value{
tibble format data table output
}
\description{
Get metadata for CoV open data dataset
}
\examples{
# Get the metadata for the street trees dataset
\dontrun{
get_cov_metadata("street-trees")
}

}
