% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{search_cov_datasets}
\alias{search_cov_datasets}
\title{Search for CoV open data datasets}
\usage{
search_cov_datasets(
  search_term,
  trim = TRUE,
  apikey = getOption("VancouverOpenDataApiKey"),
  refresh = FALSE
)
}
\arguments{
\item{search_term}{grep string to serach through datasets}

\item{trim}{trim all NA columns, optional, defaul `TRUE`}

\item{apikey}{the CoV open data API key, optional}

\item{refresh}{refresh cached data, default `FALSE``}
}
\value{
tibble format data table output
}
\description{
Search for CoV open data datasets
}
\examples{
# search available datasets relating to trees
\dontrun{
search_cov_datasets("trees")
}

}
